Refactor NavigationBar component and update NavbarLinks
