+++ frontend/app/organizations/[slug]/tournaments/[tournamentId]/pairings/page.tsx
