+++ frontend/app/organizations/[slug]/tournaments/[tournamentId]/register/page.tsx
