+++ frontend/package.json
@@ -43,6 +43,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "crypto-browserify": "^3.12.0",
+    "date-fns": "^4.1.0",
     "dotenv": "^16.4.5",
     "framer-motion": "~11.3.31",
     "intl-messageformat": "^10.5.14",
